+++ src/actions/controls.js
@@ -1,9 +1,14 @@
 import { pushToReveal } from "../utils/menu";
 import ImageData from "../models/image-data";
+import MenuLifeData from "../models/menu-life";
 import MenuRevealData from "../models/menu-reveal";
 
 export const clearReveal = (G, ctx) => {
   G.reveal[ctx.playerID] = null;
+};
+
+export const setLife = (G, ctx, lp) => {
+  G.life[ctx.playerID] = lp;
 };
 
 export const callReact = (G, ctx, image) => {
@@ -14,4 +19,10 @@ export const callReact = (G, ctx, image) => {
     MenuRevealData(topText, content),
     parseInt(ctx.playerID)
   );
+};
+
+//CLIENT
+
+export const getLifeMenu = (life) => {
+  return MenuLifeData(life);
 };