+++ src/models/menu-life.js
@@ -1,0 +1,9 @@
+const MenuLifeData = (lp) => {
+  let obj = {
+    life: lp,
+  };
+
+  return obj;
+};
+
+export default MenuLifeData;